actions: Move part into zerver.lib.subscription_info.
(cherry picked from commit 9dd7e34ab34e05b2bb51dbb480e75954974c7197)
